ISSUE-12 Allow Java 1.8 Maven Instances to Compile
Java 1.8 adds strict documentation rules as a default-enabled feature.
Currently, SPDX-Tools fails this validation. This commit adds an
argument to any maven runs where Maven is backed by Java 1.8 or newer,
which turns off the build-failure part of this check. Violations are
still printed to the build log

This has been tested against Java 1.7 and Java 1.8