LicenseExpressionParser: Allow "+" as part of license-refs
license-refs as defined in Appendix IV of the SPDX 2.0 specification
should be allowed to contain "+", e.g. to write "LicenseRef-LGPL-3.0+".
Currently, LicenseExpressionParser interprets "+" in license-refs as
Operator.OR_LATER even if they are no compound-expression. Fix this by
checking the compound-expression for a license-ref and do not take "+" as
a token in that case.